+++ Manatee.Trello.IntegrationTests/Manatee.Trello.IntegrationTests.csproj
@@ -128,8 +128,11 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ClientTests.cs" />
     <Compile Include="DevTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StackOverflowTests.cs" />
+    <Compile Include="TheGodComparer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
