+++ Manatee.Trello.IntegrationTests/StackOverflowTests.cs
@@ -1,0 +1,87 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Manatee.Trello.ManateeJson;
+using Manatee.Trello.Tests.Common;
+using Manatee.Trello.WebApi;
+using NUnit.Framework;
+
+namespace Manatee.Trello.IntegrationTests
+{
+	[TestFixture]
+	[Ignore("These tests need to work.")]
+	public class StackOverflowTests
+	{
+		private static void Run(System.Action action)
+		{
+			var serializer = new ManateeSerializer();
+			TrelloConfiguration.Serializer = serializer;
+			TrelloConfiguration.Deserializer = serializer;
+			TrelloConfiguration.JsonFactory = new ManateeFactory();
+			TrelloConfiguration.RestClientProvider = new WebApiClientProvider();
+
+			TrelloAuthorization.Default.AppKey = TrelloIds.AppKey;
+			TrelloAuthorization.Default.UserToken = TrelloIds.UserToken;
+
+			action();
+
+			TrelloProcessor.Flush();
+		}
+
+		#region http://stackoverflow.com/q/39926431/878701
+
+		private static void Move(Card card, int position, List list = null)
+		{
+			if (list != null && list != card.List)
+			{
+				card.List = list;
+			}
+
+			card.Position = position;
+		}
+
+		[Test]
+		public void MovingCards()
+		{
+			Run(() =>
+				    {
+						var list = new List(TrelloIds.ListId);
+					    var cards = new List<Card>();
+					    for (int i = 0; i < 10; i++)
+					    {
+						    cards.Add(list.Cards.Add("test card " + i));
+					    }
+
+					    var otherList = list.Board.Lists.Last();
+
+						cards.AsParallel().ForAll(c => Move(c, 1, otherList));
+
+					    foreach (var card in cards)
+					    {
+						    card.Delete();
+					    }
+					});
+		}
+
+		#endregion
+
+		#region http://stackoverflow.com/q/43667744/878701
+
+		[Test]
+		public void LabelNamesAndCards()
+		{
+			Run(() =>
+				{
+					var card = new Card(TrelloIds.CardId);
+
+					foreach (var label in card.Labels)
+					{
+						Assert.IsNotNull(label.Name);
+						// I can't always ensure Color isn't null.  Colorless labels are a thing.
+						//Assert.IsNotNull(label.Color);
+					}
+				});
+		}
+
+		#endregion
+	}
+}
