--- Manatee.Trello.Test.Unit/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="3.3.3" targetFramework="net45" />
-  <package id="Manatee.Json" version="5.1.1" targetFramework="net45" />
-  <package id="Manatee.StateMachine" version="1.1.2" targetFramework="net40" requireReinstallation="True" />
-  <package id="Moq" version="4.5.23" targetFramework="net45" />
-  <package id="StoryQ" version="2.0.5" targetFramework="net40" />
-</packages>
