--- Manatee.Trello.Test/DevTest.cs
@@ -1,54 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Manatee.Trello.CustomFields;
-using Manatee.Trello.ManateeJson;
-using Manatee.Trello.WebApi;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Manatee.Trello.Test
-{
-	[TestClass]
-	public class DevTest
-	{
-		[TestMethod]
-		public void TestMethod1()
-		{
-			Run(() =>
-				{
-					var list = new List(TrelloIds.ListId);
-					var cards = list.Cards.Filter(CardFilter.Visible);
-
-					OutputCollection("cards", cards);
-
-					TrelloProcessor.Flush();
-				});
-		}
-
-		private static void Run(System.Action action)
-		{
-			var serializer = new ManateeSerializer();
-			TrelloConfiguration.Serializer = serializer;
-			TrelloConfiguration.Deserializer = serializer;
-			TrelloConfiguration.JsonFactory = new ManateeFactory();
-			TrelloConfiguration.RestClientProvider = new WebApiClientProvider();
-
-			TrelloAuthorization.Default.AppKey = TrelloIds.AppKey;
-			TrelloAuthorization.Default.UserToken = TrelloIds.UserToken;
-
-			action();
-
-			TrelloProcessor.Flush();
-		}
-
-		private static void OutputCollection<T>(string section, IEnumerable<T> collection)
-		{
-			Console.WriteLine();
-			Console.WriteLine(section);
-			foreach (var item in collection)
-			{
-				Console.WriteLine($"    {item}");
-			}
-		}
-	}
-}
