--- Manatee.Trello.Test/app.config
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="JetBrains.Annotations" publicKeyToken="1010a0d8d6380325" culture="neutral"/>
-        <bindingRedirect oldVersion="0.0.0.0-9.2.0.0" newVersion="9.2.0.0"/>
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Manatee.StateMachine" publicKeyToken="15909d91027a225e" culture="neutral"/>
-        <bindingRedirect oldVersion="0.0.0.0-0.0.0.0" newVersion="0.0.0.0"/>
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral"/>
-        <bindingRedirect oldVersion="0.0.0.0-9.0.0.0" newVersion="9.0.0.0"/>
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6"/></startup></configuration>
