--- Manatee.Trello.Test/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Manatee.Json" version="6.0.0" targetFramework="net45" />
-  <package id="Manatee.StateMachine" version="1.1.2" targetFramework="net45" />
-  <package id="Moq" version="4.5.23" targetFramework="net45" />
-  <package id="StoryQ" version="2.0.5" targetFramework="net45" />
-</packages>
